Improved offlinelayer
